+++ apps/mini-app/src/main.ts
@@ -2,12 +2,15 @@
 
 import { inject } from "@vercel/analytics";
 import "./style.css";
-import "@fortawesome/fontawesome-free/css/fontawesome.min.css";
-import "@fortawesome/fontawesome-free/css/solid.min.css";
 import { mountApp } from "./app";
 import { APP_SHELL } from "./template";
 
-inject();
-
 document.getElementById("app")!.innerHTML = APP_SHELL;
 mountApp();
+
+const deferAnalytics = () => inject();
+if (typeof requestIdleCallback === "function") {
+  requestIdleCallback(deferAnalytics);
+} else {
+  setTimeout(deferAnalytics, 0);
+}